Room with LiveData, added isCattleExist function in cattle repository. Minor changes in updaters

+++ shared/src/main/java/com/pr656d/shared/domain/cattle/validator/CattleTagNumberValidatorUseCase.kt
@@ -53,6 +53,6 @@ class CattleTagNumberValidatorUseCase @Inject constructor(
             return
         }
 
-        isCattleExistWithTagNumberUseCase(tagNumber, isCattleExistResult)
+        isCattleExistWithTagNumberUseCase(tagNumber.toLong(), isCattleExistResult)
     }
 }
